Initial commit for index
Adding the necessary divs for the project.
Starting with the Registration form div. When registration is sucessful it will be hidden and the "home" div will appear.

I am aware there are smarter methods of manipulating the DOM without using multiple routes. I am creating this as if I had even less knowledge of Javascript.